+++ services/transaction/core/assets/registerServiceWorker.php
@@ -1,0 +1,2 @@
+<?php
+    $cid = "51ba108469bb979da044569be9a4c3e0798b836e6f07747b310c4c2cfe012b55";
